+++ source/renderer/app/domains/Wallet.js
@@ -39,6 +39,51 @@ export const WalletDelegationStatuses: {
   NOT_DELEGATING: 'not_delegating',
 };
 
+export type HwDeviceStatus =
+  | 'connecting'
+  | 'connecting_failed'
+  | 'trezor_bridge_failure'
+  | 'launching_cardano_app'
+  | 'exporting_public_key'
+  | 'exporting_public_key_failed'
+  | 'ready'
+  | 'verifying_transaction'
+  | 'verifying_transaction_failed'
+  | 'wrong_firmware'
+  | 'wrong_cardano_app_version'
+  | 'unsupported_device'
+  | 'verifying_transaction_succeeded';
+
+export const HwDeviceStatuses: {
+  CONNECTING: HwDeviceStatus,
+  CONNECTING_FAILED: HwDeviceStatus,
+  LAUNCHING_CARDANO_APP: HwDeviceStatus,
+  EXPORTING_PUBLIC_KEY: HwDeviceStatus,
+  EXPORTING_PUBLIC_KEY_FAILED: HwDeviceStatus,
+  READY: HwDeviceStatus,
+  VERIFYING_TRANSACTION: HwDeviceStatus,
+  VERIFYING_TRANSACTION_FAILED: HwDeviceStatus,
+  VERIFYING_TRANSACTION_SUCCEEDED: HwDeviceStatus,
+  WRONG_FIRMWARE: HwDeviceStatus,
+  WRONG_CARDANO_APP_VERSION: HwDeviceStatus,
+  UNSUPPORTED_DEVICE: HwDeviceStatus,
+  TREZOR_BRIDGE_FAILURE: HwDeviceStatus,
+} = {
+  CONNECTING: 'connecting',
+  CONNECTING_FAILED: 'connecting_failed',
+  TREZOR_BRIDGE_FAILURE: 'trezor_bridge_failure',
+  LAUNCHING_CARDANO_APP: 'launching_cardano_app',
+  EXPORTING_PUBLIC_KEY: 'exporting_public_key',
+  EXPORTING_PUBLIC_KEY_FAILED: 'exporting_public_key_failed',
+  WRONG_FIRMWARE: 'wrong_firmware',
+  WRONG_CARDANO_APP_VERSION: 'wrong_cardano_app_version',
+  UNSUPPORTED_DEVICE: 'unsupported_device',
+  READY: 'ready',
+  VERIFYING_TRANSACTION: 'verifying_transaction',
+  VERIFYING_TRANSACTION_FAILED: 'verifying_transaction_failed',
+  VERIFYING_TRANSACTION_SUCCEEDED: 'verifying_transaction_succeeded',
+};
+
 export const WalletUnits: {
   ADA: WalletUnit,
   LOVELACE: WalletUnit,
@@ -83,6 +128,7 @@ export default class Wallet {
   @observable discovery: Discovery;
   @observable hasPassword: boolean;
   @observable walletNotConnected: boolean;
+  @observable isHardwareWallet: boolean;
 
   constructor(data: WalletProps) {
     Object.assign(this, data);
@@ -108,6 +154,7 @@ export default class Wallet {
         'discovery',
         'hasPassword',
         'walletNotConnected',
+        'isHardwareWallet',
       ])
     );
   }